move httpclient extensions to Nox.Core
Add extension to set etga
