+++ src/Nox.Core/Application/Dto/ODataSingleDto.cs
@@ -1,8 +1,8 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace ClientApi.Tests.Tests.Models;
+namespace Nox.Application.Dto;
 
-public class ODataSigleResponse
+public class ODataSingleDto
 {
     [JsonPropertyName("@odata.context")]
     public string Context { get; set; } = default!;
