+++ src/Nox.Core/Extensions/HttpClientExtensions.cs
@@ -1,20 +1,24 @@
 ﻿using System.Text.Json.Serialization;
 using System.Text.Json;
-using Nox.Ui.Blazor.Lib.Models;
+
 using System.Net.Http.Json;
+using Nox.Application.Dto;
+using System.Net.Http.Headers;
 
-namespace Nox.Ui.Blazor.Lib.Extensions;
+namespace Nox.Extensions;
 
-public static class HttpClientExtension
+public static class HttpClientExtensions
 {
     public static async Task<TResult?> GetODataCollectionResponseAsync<TResult>(this HttpClient httpClient, string requestUrl)
     {
+      
         var result = await httpClient.GetAsync(requestUrl);
+
         result.EnsureSuccessStatusCode();
 
         var content = await result.Content.ReadAsStringAsync();
-        var data = DeserializeResponse<ODataCollectionResponse<TResult>>(content);
-        
+        var data = DeserializeResponse<ODataCollectionDto<TResult>>(content);
+
         return data!.Value;
     }
 
@@ -35,7 +39,8 @@ public static class HttpClientExtension
         var content = await result.Content.ReadAsStringAsync();
 
         return DeserializeResponse<TResult>(content);
-    }
+    }    
+
     public static async Task<TResult?> PutAsync<TValue, TResult>(this HttpClient httpClient, string requestUrl, TValue data)
     {
         var result = await httpClient.PutAsJsonAsync(requestUrl, data);
@@ -45,9 +50,23 @@ public static class HttpClientExtension
 
         return DeserializeResponse<TResult>(content);
     }
+    public static async Task<HttpResponseMessage?> DeleteAsync(this HttpClient httpClient, string requestUrl)
+    {
+        var message = await httpClient.DeleteAsync(requestUrl);
+
+         message.EnsureSuccessStatusCode();
+
+        return message;
+    }
+
+    public static void AddeTag(this HttpClient httpClient, Guid eTag)
+    {
+        httpClient.DefaultRequestHeaders.IfNoneMatch.Add(new EntityTagHeaderValue(eTag.ToString()));
+    }
 
     private static TResult? DeserializeResponse<TResult>(string response)
     {
         return JsonSerializer.Deserialize<TResult>(response, new JsonSerializerOptions { PropertyNameCaseInsensitive = true, Converters = { new JsonStringEnumConverter() } });
     }
+    
 }
