--- tests/Nox.ClientApi.Tests/Tests/Models/ODataCollectionResponse.cs
@@ -1,11 +1,0 @@
-﻿using System.Text.Json.Serialization;
-
-namespace ClientApi.Tests.Tests.Models;
-
-public class ODataCollectionResponse<T>
-{
-    public T Value { get; set; } = default!;
-    
-    [JsonPropertyName("@odata.count")]
-    public int Count { get; set; }
-}
